Backend/requirements.txt
flask
flask-cors
numpy
pillow
requests
tensorflow-cpu
gunicorn